Add a rule that links accounts together during login
When a user logs in, this checks to see if any other users with the same email address exist (excluding the account that will just have been created). If so it merges the accounts together to create a new account with the current connection being the primary identity of that user.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>