Add a rule that injects the ID from the app_metadata into the idToken
The client application needs to know the ID of the associated user object stored on our database (which we store in the app_metadata for the Auth0 user). As we cannot request this information in the scope (see https://auth0.com/docs/api-auth/tutorials/adoption/scope-custom-claims) we need to use a rule to inject this information into the token.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>